Reorganize stanfit into smaller files
