+++ cmdstanpy/stanfit/mcmc.py
@@ -1,15 +1,12 @@
-"""Container objects for results of CmdStan run(s)."""
+"""
+Container for the result of running the sample (MCMC) method
+or generate quantities (GQ) method
+"""
 
-import copy
-import glob
 import math
 import os
-import re
-import shutil
-from collections import Counter, OrderedDict
-from datetime import datetime
+from collections import Counter
 from io import StringIO
-from time import time
 from typing import (
     Any,
     Dict,
@@ -32,13 +29,7 @@ except ImportError:
     XARRAY_INSTALLED = False
 
 from cmdstanpy import _CMDSTAN_SAMPLING, _CMDSTAN_THIN, _CMDSTAN_WARMUP, _TMPDIR
-from cmdstanpy.cmdstan_args import (
-    CmdStanArgs,
-    Method,
-    OptimizeArgs,
-    SamplerArgs,
-    VariationalArgs,
-)
+from cmdstanpy.cmdstan_args import Method, SamplerArgs
 from cmdstanpy.utils import (
     EXTENSION,
     check_sampler_csv,
@@ -48,352 +39,11 @@ from cmdstanpy.utils import (
     do_command,
     flatten_chains,
     get_logger,
-    parse_method_vars,
-    parse_stan_vars,
-    scan_config,
     scan_generated_quantities_csv,
-    scan_optimize_csv,
-    scan_variational_csv,
 )
 
-
-class RunSet:
-    """
-    Encapsulates the configuration and results of a call to any CmdStan
-    inference method. Records the method return code and locations of
-    all console, error, and output files.
-
-    RunSet objects are instantiated by the CmdStanModel class inference methods
-    which validate all inputs, therefore "__init__" method skips input checks.
-    """
-
-    def __init__(
-        self,
-        args: CmdStanArgs,
-        chains: int = 1,
-        *,
-        chain_ids: Optional[List[int]] = None,
-        time_fmt: str = "%Y%m%d%H%M%S",
-        one_process_per_chain: bool = True,
-    ) -> None:
-        """Initialize object (no input arg checks)."""
-        self._args = args
-        self._chains = chains
-        self._one_process_per_chain = one_process_per_chain
-        if one_process_per_chain:
-            self._num_procs = chains
-        else:
-            self._num_procs = 1
-        self._retcodes = [-1 for _ in range(self._num_procs)]
-        if chain_ids is None:
-            chain_ids = [i + 1 for i in range(chains)]
-        self._chain_ids = chain_ids
-
-        if args.output_dir is not None:
-            self._output_dir = args.output_dir
-        else:
-            self._output_dir = _TMPDIR
-
-        # output files prefix: ``<model_name>-<YYYYMMDDHHMM>_<chain_id>``
-        self._base_outfile = (
-            f'{args.model_name}-{datetime.now().strftime(time_fmt)}'
-        )
-        # per-process console messages
-        self._stdout_files = [''] * self._num_procs
-        if one_process_per_chain:
-            for i in range(chains):
-                self._stdout_files[i] = self.file_path("-stdout.txt", id=i)
-        else:
-            self._stdout_files[0] = self.file_path("-stdout.txt")
-
-        # per-chain output files
-        self._csv_files = [''] * chains
-        self._diagnostic_files = [''] * chains  # optional
-        self._profile_files = [''] * chains  # optional
-
-        if chains == 1:
-            self._csv_files[0] = self.file_path(".csv")
-            if args.save_latent_dynamics:
-                self._diagnostic_files[0] = self.file_path(
-                    ".csv", extra="-diagnostic"
-                )
-            if args.save_profile:
-                self._profile_files[0] = self.file_path(
-                    ".csv", extra="-profile"
-                )
-        else:
-            for i in range(chains):
-                self._csv_files[i] = self.file_path(".csv", id=chain_ids[i])
-                if args.save_latent_dynamics:
-                    self._diagnostic_files[i] = self.file_path(
-                        ".csv", extra="-diagnostic", id=chain_ids[i]
-                    )
-                if args.save_profile:
-                    self._profile_files[i] = self.file_path(
-                        ".csv", extra="-profile", id=chain_ids[i]
-                    )
-
-    def __repr__(self) -> str:
-        repr = 'RunSet: chains={}, chain_ids={}, num_processes={}'.format(
-            self._chains, self._chain_ids, self._num_procs
-        )
-        repr = '{}\n cmd (chain 1):\n\t{}'.format(repr, self.cmd(0))
-        repr = '{}\n retcodes={}'.format(repr, self._retcodes)
-        repr = f'{repr}\n per-chain output files (showing chain 1 only):'
-        repr = '{}\n csv_file:\n\t{}'.format(repr, self._csv_files[0])
-        if self._args.save_latent_dynamics:
-            repr = '{}\n diagnostics_file:\n\t{}'.format(
-                repr, self._diagnostic_files[0]
-            )
-        if self._args.save_profile:
-            repr = '{}\n profile_file:\n\t{}'.format(
-                repr, self._profile_files[0]
-            )
-        repr = '{}\n console_msgs (if any):\n\t{}'.format(
-            repr, self._stdout_files[0]
-        )
-        return repr
-
-    @property
-    def model(self) -> str:
-        """Stan model name."""
-        return self._args.model_name
-
-    @property
-    def method(self) -> Method:
-        """CmdStan method used to generate this fit."""
-        return self._args.method
-
-    @property
-    def num_procs(self) -> int:
-        """Number of processes run."""
-        return self._num_procs
-
-    @property
-    def one_process_per_chain(self) -> bool:
-        """
-        When True, for each chain, call CmdStan in its own subprocess.
-        When False, use CmdStan's `num_chains` arg to run parallel chains.
-        Always True if CmdStan < 2.28.
-        For CmdStan 2.28 and up, `sample` method determines value.
-        """
-        return self._one_process_per_chain
-
-    @property
-    def chains(self) -> int:
-        """Number of chains."""
-        return self._chains
-
-    @property
-    def chain_ids(self) -> List[int]:
-        """Chain ids."""
-        return self._chain_ids
-
-    def cmd(self, idx: int) -> List[str]:
-        """
-        Assemble CmdStan invocation.
-        When running parallel chains from single process (2.28 and up),
-        specify CmdStan arg `num_chains` and leave chain idx off CSV files.
-        """
-        if self._one_process_per_chain:
-            return self._args.compose_command(
-                idx,
-                csv_file=self.csv_files[idx],
-                diagnostic_file=self.diagnostic_files[idx]
-                if self._args.save_latent_dynamics
-                else None,
-                profile_file=self.profile_files[idx]
-                if self._args.save_profile
-                else None,
-            )
-        else:
-            return self._args.compose_command(
-                idx,
-                csv_file=self.file_path('.csv'),
-                diagnostic_file=self.file_path(".csv", extra="-diagnostic")
-                if self._args.save_latent_dynamics
-                else None,
-                profile_file=self.file_path(".csv", extra="-profile")
-                if self._args.save_profile
-                else None,
-                num_chains=self._chains,
-            )
-
-    @property
-    def csv_files(self) -> List[str]:
-        """List of paths to CmdStan output files."""
-        return self._csv_files
-
-    @property
-    def stdout_files(self) -> List[str]:
-        """
-        List of paths to transcript of CmdStan messages sent to the console.
-        Transcripts include config information, progress, and error messages.
-        """
-        return self._stdout_files
-
-    def _check_retcodes(self) -> bool:
-        """Returns ``True`` when all chains have retcode 0."""
-        for code in self._retcodes:
-            if code != 0:
-                return False
-        return True
-
-    @property
-    def diagnostic_files(self) -> List[str]:
-        """List of paths to CmdStan hamiltonian diagnostic files."""
-        return self._diagnostic_files
-
-    @property
-    def profile_files(self) -> List[str]:
-        """List of paths to CmdStan profiler files."""
-        return self._profile_files
-
-    # pylint: disable=invalid-name
-    def file_path(
-        self, suffix: str, *, extra: str = "", id: Optional[int] = None
-    ) -> str:
-        if id is not None:
-            suffix = f"_{id}{suffix}"
-        file = os.path.join(
-            self._output_dir, f"{self._base_outfile}{extra}{suffix}"
-        )
-        return file
-
-    def _retcode(self, idx: int) -> int:
-        """Get retcode for process[idx]."""
-        return self._retcodes[idx]
-
-    def _set_retcode(self, idx: int, val: int) -> None:
-        """Set retcode at process[idx] to val."""
-        self._retcodes[idx] = val
-
-    def get_err_msgs(self) -> str:
-        """Checks console messages for each CmdStan run."""
-        msgs = []
-        for i in range(self._num_procs):
-            if (
-                os.path.exists(self._stdout_files[i])
-                and os.stat(self._stdout_files[i]).st_size > 0
-            ):
-                if self._args.method == Method.OPTIMIZE:
-                    msgs.append('console log output:\n')
-                    with open(self._stdout_files[0], 'r') as fd:
-                        msgs.append(fd.read())
-                else:
-                    with open(self._stdout_files[i], 'r') as fd:
-                        contents = fd.read()
-                        # pattern matches initial "Exception" or "Error" msg
-                        pat = re.compile(r'^E[rx].*$', re.M)
-                        errors = re.findall(pat, contents)
-                        if len(errors) > 0:
-                            msgs.append('\n\t'.join(errors))
-        return '\n'.join(msgs)
-
-    def save_csvfiles(self, dir: Optional[str] = None) -> None:
-        """
-        Moves CSV files to specified directory.
-
-        :param dir: directory path
-
-        See Also
-        --------
-        cmdstanpy.from_csv
-        """
-        if dir is None:
-            dir = os.path.realpath('.')
-        test_path = os.path.join(dir, str(time()))
-        try:
-            os.makedirs(dir, exist_ok=True)
-            with open(test_path, 'w'):
-                pass
-            os.remove(test_path)  # cleanup
-        except (IOError, OSError, PermissionError) as exc:
-            raise Exception('Cannot save to path: {}'.format(dir)) from exc
-
-        for i in range(self.chains):
-            if not os.path.exists(self._csv_files[i]):
-                raise ValueError(
-                    'Cannot access CSV file {}'.format(self._csv_files[i])
-                )
-
-            to_path = os.path.join(dir, os.path.basename(self._csv_files[i]))
-            if os.path.exists(to_path):
-                raise ValueError(
-                    'File exists, not overwriting: {}'.format(to_path)
-                )
-            try:
-                get_logger().debug(
-                    'saving tmpfile: "%s" as: "%s"', self._csv_files[i], to_path
-                )
-                shutil.move(self._csv_files[i], to_path)
-                self._csv_files[i] = to_path
-            except (IOError, OSError, PermissionError) as e:
-                raise ValueError(
-                    'Cannot save to file: {}'.format(to_path)
-                ) from e
-
-
-class InferenceMetadata:
-    """
-    CmdStan configuration and contents of output file parsed out of
-    the Stan CSV file header comments and column headers.
-    Assumes valid CSV files.
-    """
-
-    def __init__(self, config: Dict[str, Any]) -> None:
-        """Initialize object from CSV headers"""
-        self._cmdstan_config = config
-        self._method_vars_cols = parse_method_vars(names=config['column_names'])
-        stan_vars_dims, stan_vars_cols = parse_stan_vars(
-            names=config['column_names']
-        )
-        self._stan_vars_dims = stan_vars_dims
-        self._stan_vars_cols = stan_vars_cols
-
-    def __repr__(self) -> str:
-        return 'Metadata:\n{}\n'.format(self._cmdstan_config)
-
-    @property
-    def cmdstan_config(self) -> Dict[str, Any]:
-        """
-        Returns a dictionary containing a set of name, value pairs
-        parsed out of the Stan CSV file header.  These include the
-        command configuration and the CSV file header row information.
-        Uses deepcopy for immutability.
-        """
-        return copy.deepcopy(self._cmdstan_config)
-
-    @property
-    def method_vars_cols(self) -> Dict[str, Tuple[int, ...]]:
-        """
-        Returns a map from a Stan inference method variable to
-        a tuple of column indices in inference engine's output array.
-        Method variable names always end in `__`, e.g. `lp__`.
-        Uses deepcopy for immutability.
-        """
-        return copy.deepcopy(self._method_vars_cols)
-
-    @property
-    def stan_vars_cols(self) -> Dict[str, Tuple[int, ...]]:
-        """
-        Returns a map from a Stan program variable name to a
-        tuple of the column indices in the vector or matrix of
-        estimates produced by a CmdStan inference method.
-        Uses deepcopy for immutability.
-        """
-        return copy.deepcopy(self._stan_vars_cols)
-
-    @property
-    def stan_vars_dims(self) -> Dict[str, Tuple[int, ...]]:
-        """
-        Returns map from Stan program variable names to variable dimensions.
-        Scalar types are mapped to the empty tuple, e.g.,
-        program variable ``int foo`` has dimension ``()`` and
-        program variable ``vector[10] bar`` has single dimension ``(10)``.
-        Uses deepcopy for immutability.
-        """
-        return copy.deepcopy(self._stan_vars_dims)
+from .metadata import InferenceMetadata
+from .runset import RunSet
 
 
 class CmdStanMCMC:
@@ -1078,263 +728,6 @@ class CmdStanMCMC:
         self.runset.save_csvfiles(dir)
 
 
-class CmdStanMLE:
-    """
-    Container for outputs from CmdStan optimization.
-    Created by :meth:`CmdStanModel.optimize`.
-    """
-
-    def __init__(self, runset: RunSet) -> None:
-        """Initialize object."""
-        if not runset.method == Method.OPTIMIZE:
-            raise ValueError(
-                'Wrong runset method, expecting optimize runset, '
-                'found method {}'.format(runset.method)
-            )
-        self.runset = runset
-        # info from runset to be exposed
-        self.converged = runset._check_retcodes()
-        optimize_args = self.runset._args.method_args
-        assert isinstance(
-            optimize_args, OptimizeArgs
-        )  # make the typechecker happy
-        self._save_iterations = optimize_args.save_iterations
-        self._set_mle_attrs(runset.csv_files[0])
-
-    def __repr__(self) -> str:
-        repr = 'CmdStanMLE: model={}{}'.format(
-            self.runset.model, self.runset._args.method_args.compose(0, cmd=[])
-        )
-        repr = '{}\n csv_file:\n\t{}\n output_file:\n\t{}'.format(
-            repr,
-            '\n\t'.join(self.runset.csv_files),
-            '\n\t'.join(self.runset.stdout_files),
-        )
-        if not self.converged:
-            repr = '{}\n Warning: invalid estimate, '.format(repr)
-            repr = '{} optimization failed to converge.'.format(repr)
-        return repr
-
-    def _set_mle_attrs(self, sample_csv_0: str) -> None:
-        meta = scan_optimize_csv(sample_csv_0, self._save_iterations)
-        self._metadata = InferenceMetadata(meta)
-        self._column_names: Tuple[str, ...] = meta['column_names']
-        assert isinstance(meta['mle'], np.ndarray)  # make the typechecker happy
-        self._mle = meta['mle']
-        if self._save_iterations:
-            assert isinstance(
-                meta['all_iters'], np.ndarray
-            )  # make the typechecker happy
-            self._all_iters = meta['all_iters']
-
-    @property
-    def column_names(self) -> Tuple[str, ...]:
-        """
-        Names of estimated quantities, includes joint log probability,
-        and all parameters, transformed parameters, and generated quantities.
-        """
-        return self._column_names
-
-    @property
-    def metadata(self) -> InferenceMetadata:
-        """
-        Returns object which contains CmdStan configuration as well as
-        information about the names and structure of the inference method
-        and model output variables.
-        """
-        return self._metadata
-
-    @property
-    def optimized_params_np(self) -> np.ndarray:
-        """
-        Returns all final estimates from the optimizer as a numpy.ndarray
-        which contains all optimizer outputs, i.e., the value for `lp__`
-        as well as all Stan program variables.
-        """
-        if not self.converged:
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-        return self._mle
-
-    @property
-    def optimized_iterations_np(self) -> Optional[np.ndarray]:
-        """
-        Returns all saved iterations from the optimizer and final estimate
-        as a numpy.ndarray which contains all optimizer outputs, i.e.,
-        the value for `lp__` as well as all Stan program variables.
-
-        """
-        if not self._save_iterations:
-            get_logger().warning(
-                'Intermediate iterations not saved to CSV output file. '
-                'Rerun the optimize method with "save_iterations=True".'
-            )
-            return None
-        if not self.converged:
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-        return self._all_iters
-
-    @property
-    def optimized_params_pd(self) -> pd.DataFrame:
-        """
-        Returns all final estimates from the optimizer as a pandas.DataFrame
-        which contains all optimizer outputs, i.e., the value for `lp__`
-        as well as all Stan program variables.
-        """
-        if not self.runset._check_retcodes():
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-        return pd.DataFrame([self._mle], columns=self.column_names)
-
-    @property
-    def optimized_iterations_pd(self) -> Optional[pd.DataFrame]:
-        """
-        Returns all saved iterations from the optimizer and final estimate
-        as a pandas.DataFrame which contains all optimizer outputs, i.e.,
-        the value for `lp__` as well as all Stan program variables.
-
-        """
-        if not self._save_iterations:
-            get_logger().warning(
-                'Intermediate iterations not saved to CSV output file. '
-                'Rerun the optimize method with "save_iterations=True".'
-            )
-            return None
-        if not self.converged:
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-        return pd.DataFrame(self._all_iters, columns=self.column_names)
-
-    @property
-    def optimized_params_dict(self) -> Dict[str, float]:
-        """
-        Returns all estimates from the optimizer, including `lp__` as a
-        Python Dict.  Only returns estimate from final iteration.
-        """
-        if not self.runset._check_retcodes():
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-        return OrderedDict(zip(self.column_names, self._mle))
-
-    def stan_variable(
-        self,
-        var: Optional[str] = None,
-        *,
-        inc_iterations: bool = False,
-        warn: bool = True,
-    ) -> Union[np.ndarray, float]:
-        """
-        Return a numpy.ndarray which contains the estimates for the
-        for the named Stan program variable where the dimensions of the
-        numpy.ndarray match the shape of the Stan program variable.
-
-        :param var: variable name
-
-        :param inc_iterations: When ``True`` and the intermediate estimates
-            are included in the output, i.e., the optimizer was run with
-            ``save_iterations=True``, then intermediate estimates are included.
-            Default value is ``False``.
-
-        See Also
-        --------
-        CmdStanMLE.stan_variables
-        CmdStanMCMC.stan_variable
-        CmdStanVB.stan_variable
-        CmdStanGQ.stan_variable
-        """
-        if var is None:
-            raise ValueError('no variable name specified.')
-        if var not in self._metadata.stan_vars_dims:
-            raise ValueError('unknown variable name: {}'.format(var))
-        if warn and inc_iterations and not self._save_iterations:
-            get_logger().warning(
-                'Intermediate iterations not saved to CSV output file. '
-                'Rerun the optimize method with "save_iterations=True".'
-            )
-        if warn and not self.runset._check_retcodes():
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-
-        col_idxs = list(self._metadata.stan_vars_cols[var])
-        if inc_iterations and self._save_iterations:
-            num_rows = self._all_iters.shape[0]
-        else:
-            num_rows = 1
-        if len(col_idxs) > 1:  # container var
-            dims = (num_rows,) + self._metadata.stan_vars_dims[var]
-            # pylint: disable=redundant-keyword-arg
-            if num_rows > 1:
-                result = self._all_iters[:, col_idxs].reshape(  # type: ignore
-                    dims, order='F'
-                )
-            else:
-                result = self._mle[col_idxs].reshape(dims[1:], order="F")
-        else:  # scalar var
-            col_idx = col_idxs[0]
-            if num_rows > 1:
-                result = self._all_iters[:, col_idx]
-            else:
-                result = float(self._mle[col_idx])
-        assert isinstance(
-            result, (np.ndarray, float)
-        )  # make the typechecker happy
-        return result
-
-    def stan_variables(
-        self, inc_iterations: bool = False
-    ) -> Dict[str, Union[np.ndarray, float]]:
-        """
-        Return a dictionary mapping Stan program variables names
-        to the corresponding numpy.ndarray containing the inferred values.
-
-        :param inc_iterations: When ``True`` and the intermediate estimates
-            are included in the output, i.e., the optimizer was run with
-            ``save_iterations=True``, then intermediate estimates are included.
-            Default value is ``False``.
-
-
-        See Also
-        --------
-        CmdStanMLE.stan_variable
-        CmdStanMCMC.stan_variables
-        CmdStanVB.stan_variables
-        CmdStanGQ.stan_variables
-        """
-        if not self.runset._check_retcodes():
-            get_logger().warning(
-                'Invalid estimate, optimization failed to converge.'
-            )
-        result = {}
-        for name in self._metadata.stan_vars_dims.keys():
-            result[name] = self.stan_variable(
-                name, inc_iterations=inc_iterations, warn=False
-            )
-        return result
-
-    def save_csvfiles(self, dir: Optional[str] = None) -> None:
-        """
-        Move output CSV files to specified directory.  If files were
-        written to the temporary session directory, clean filename.
-        E.g., save 'bernoulli-201912081451-1-5nm6as7u.csv' as
-        'bernoulli-201912081451-1.csv'.
-
-        :param dir: directory path
-
-        See Also
-        --------
-        stanfit.RunSet.save_csvfiles
-        cmdstanpy.from_csv
-        """
-        self.runset.save_csvfiles(dir)
-
-
 class CmdStanGQ:
     """
     Container for outputs from CmdStan generate_quantities run.
@@ -1860,359 +1253,6 @@ class CmdStanGQ:
         cmdstanpy.from_csv
         """
         self.runset.save_csvfiles(dir)
-
-
-class CmdStanVB:
-    """
-    Container for outputs from CmdStan variational run.
-    Created by :meth:`CmdStanModel.variational`.
-    """
-
-    def __init__(self, runset: RunSet) -> None:
-        """Initialize object."""
-        if not runset.method == Method.VARIATIONAL:
-            raise ValueError(
-                'Wrong runset method, expecting variational inference, '
-                'found method {}'.format(runset.method)
-            )
-        self.runset = runset
-        self._set_variational_attrs(runset.csv_files[0])
-
-    def __repr__(self) -> str:
-        repr = 'CmdStanVB: model={}{}'.format(
-            self.runset.model, self.runset._args.method_args.compose(0, cmd=[])
-        )
-        repr = '{}\n csv_file:\n\t{}\n output_file:\n\t{}'.format(
-            repr,
-            '\n\t'.join(self.runset.csv_files),
-            '\n\t'.join(self.runset.stdout_files),
-        )
-        # TODO - diagnostic, profiling files
-        return repr
-
-    def _set_variational_attrs(self, sample_csv_0: str) -> None:
-        meta = scan_variational_csv(sample_csv_0)
-        self._metadata = InferenceMetadata(meta)
-        # these three assignments don't grant type information
-        self._column_names: Tuple[str, ...] = meta['column_names']
-        self._eta: float = meta['eta']
-        self._variational_mean: np.ndarray = meta['variational_mean']
-        self._variational_sample: np.ndarray = meta['variational_sample']
-
-    @property
-    def columns(self) -> int:
-        """
-        Total number of information items returned by sampler.
-        Includes approximation information and names of model parameters
-        and computed quantities.
-        """
-        return len(self._column_names)
-
-    @property
-    def column_names(self) -> Tuple[str, ...]:
-        """
-        Names of information items returned by sampler for each draw.
-        Includes approximation information and names of model parameters
-        and computed quantities.
-        """
-        return self._column_names
-
-    @property
-    def eta(self) -> float:
-        """
-        Step size scaling parameter 'eta'
-        """
-        return self._eta
-
-    @property
-    def variational_params_np(self) -> np.ndarray:
-        """
-        Returns inferred parameter means as numpy array.
-        """
-        return self._variational_mean
-
-    @property
-    def variational_params_pd(self) -> pd.DataFrame:
-        """
-        Returns inferred parameter means as pandas DataFrame.
-        """
-        return pd.DataFrame([self._variational_mean], columns=self.column_names)
-
-    @property
-    def variational_params_dict(self) -> Dict[str, np.ndarray]:
-        """Returns inferred parameter means as Dict."""
-        return OrderedDict(zip(self.column_names, self._variational_mean))
-
-    @property
-    def metadata(self) -> InferenceMetadata:
-        """
-        Returns object which contains CmdStan configuration as well as
-        information about the names and structure of the inference method
-        and model output variables.
-        """
-        return self._metadata
-
-    def stan_variable(
-        self, var: Optional[str] = None
-    ) -> Union[np.ndarray, float]:
-        """
-        Return a numpy.ndarray which contains the estimates for the
-        for the named Stan program variable where the dimensions of the
-        numpy.ndarray match the shape of the Stan program variable.
-
-        :param var: variable name
-
-        See Also
-        --------
-        CmdStanVB.stan_variables
-        CmdStanMCMC.stan_variable
-        CmdStanMLE.stan_variable
-        CmdStanGQ.stan_variable
-        """
-        if var is None:
-            raise ValueError('No variable name specified.')
-        if var not in self._metadata.stan_vars_dims:
-            raise ValueError('Unknown variable name: {}'.format(var))
-        col_idxs = list(self._metadata.stan_vars_cols[var])
-        shape: Tuple[int, ...] = ()
-        if len(col_idxs) > 1:
-            shape = self._metadata.stan_vars_dims[var]
-            result = np.asarray(self._variational_mean)[col_idxs].reshape(
-                shape, order="F"
-            )
-        else:
-            result = float(self._variational_mean[col_idxs[0]])
-        assert isinstance(result, (np.ndarray, float))
-        return result
-
-    def stan_variables(self) -> Dict[str, Union[np.ndarray, float]]:
-        """
-        Return a dictionary mapping Stan program variables names
-        to the corresponding numpy.ndarray containing the inferred values.
-
-        See Also
-        --------
-        CmdStanVB.stan_variable
-        CmdStanMCMC.stan_variables
-        CmdStanMLE.stan_variables
-        CmdStanGQ.stan_variables
-        """
-        result = {}
-        for name in self._metadata.stan_vars_dims.keys():
-            result[name] = self.stan_variable(name)
-        return result
-
-    @property
-    def variational_sample(self) -> np.ndarray:
-        """Returns the set of approximate posterior output draws."""
-        return self._variational_sample
-
-    def save_csvfiles(self, dir: Optional[str] = None) -> None:
-        """
-        Move output CSV files to specified directory.  If files were
-        written to the temporary session directory, clean filename.
-        E.g., save 'bernoulli-201912081451-1-5nm6as7u.csv' as
-        'bernoulli-201912081451-1.csv'.
-
-        :param dir: directory path
-
-        See Also
-        --------
-        stanfit.RunSet.save_csvfiles
-        cmdstanpy.from_csv
-        """
-        self.runset.save_csvfiles(dir)
-
-
-def from_csv(
-    path: Union[str, List[str], None] = None, method: Optional[str] = None
-) -> Union[CmdStanMCMC, CmdStanMLE, CmdStanVB, None]:
-    """
-    Instantiate a CmdStan object from a the Stan CSV files from a CmdStan run.
-    CSV files are specified from either a list of Stan CSV files or a single
-    filepath which can be either a directory name, a Stan CSV filename, or
-    a pathname pattern (i.e., a Python glob).  The optional argument 'method'
-    checks that the CSV files were produced by that method.
-    Stan CSV files from CmdStan methods 'sample', 'optimize', and 'variational'
-    result in objects of class CmdStanMCMC, CmdStanMLE, and CmdStanVB,
-    respectively.
-
-    :param path: directory path
-    :param method: method name (optional)
-
-    :return: either a CmdStanMCMC, CmdStanMLE, or CmdStanVB object
-    """
-    if path is None:
-        raise ValueError('Must specify path to Stan CSV files.')
-    if method is not None and method not in [
-        'sample',
-        'optimize',
-        'variational',
-    ]:
-        raise ValueError(
-            'Bad method argument {}, must be one of: '
-            '"sample", "optimize", "variational"'.format(method)
-        )
-
-    csvfiles = []
-    if isinstance(path, list):
-        csvfiles = path
-    elif isinstance(path, str):
-        if '*' in path:
-            splits = os.path.split(path)
-            if splits[0] is not None:
-                if not (os.path.exists(splits[0]) and os.path.isdir(splits[0])):
-                    raise ValueError(
-                        'Invalid path specification, {} '
-                        ' unknown directory: {}'.format(path, splits[0])
-                    )
-            csvfiles = glob.glob(path)
-        elif os.path.exists(path) and os.path.isdir(path):
-            for file in os.listdir(path):
-                if file.endswith(".csv"):
-                    csvfiles.append(os.path.join(path, file))
-        elif os.path.exists(path):
-            csvfiles.append(path)
-        else:
-            raise ValueError('Invalid path specification: {}'.format(path))
-    else:
-        raise ValueError('Invalid path specification: {}'.format(path))
-
-    if len(csvfiles) == 0:
-        raise ValueError('No CSV files found in directory {}'.format(path))
-    for file in csvfiles:
-        if not (os.path.exists(file) and file.endswith('.csv')):
-            raise ValueError(
-                'Bad CSV file path spec,'
-                ' includes non-csv file: {}'.format(file)
-            )
-
-    config_dict: Dict[str, Any] = {}
-    try:
-        with open(csvfiles[0], 'r') as fd:
-            scan_config(fd, config_dict, 0)
-    except (IOError, OSError, PermissionError) as e:
-        raise ValueError('Cannot read CSV file: {}'.format(csvfiles[0])) from e
-    if 'model' not in config_dict or 'method' not in config_dict:
-        raise ValueError("File {} is not a Stan CSV file.".format(csvfiles[0]))
-    if method is not None and method != config_dict['method']:
-        raise ValueError(
-            'Expecting Stan CSV output files from method {}, '
-            ' found outputs from method {}'.format(
-                method, config_dict['method']
-            )
-        )
-    try:
-        if config_dict['method'] == 'sample':
-            chains = len(csvfiles)
-            sampler_args = SamplerArgs(
-                iter_sampling=config_dict['num_samples'],
-                iter_warmup=config_dict['num_warmup'],
-                thin=config_dict['thin'],
-                save_warmup=config_dict['save_warmup'],
-            )
-            # bugfix 425, check for fixed_params output
-            try:
-                check_sampler_csv(
-                    csvfiles[0],
-                    iter_sampling=config_dict['num_samples'],
-                    iter_warmup=config_dict['num_warmup'],
-                    thin=config_dict['thin'],
-                    save_warmup=config_dict['save_warmup'],
-                )
-            except ValueError:
-                try:
-                    check_sampler_csv(
-                        csvfiles[0],
-                        is_fixed_param=True,
-                        iter_sampling=config_dict['num_samples'],
-                        iter_warmup=config_dict['num_warmup'],
-                        thin=config_dict['thin'],
-                        save_warmup=config_dict['save_warmup'],
-                    )
-                    sampler_args = SamplerArgs(
-                        iter_sampling=config_dict['num_samples'],
-                        iter_warmup=config_dict['num_warmup'],
-                        thin=config_dict['thin'],
-                        save_warmup=config_dict['save_warmup'],
-                        fixed_param=True,
-                    )
-                except (ValueError) as e:
-                    raise ValueError(
-                        'Invalid or corrupt Stan CSV output file, '
-                    ) from e
-
-            cmdstan_args = CmdStanArgs(
-                model_name=config_dict['model'],
-                model_exe=config_dict['model'],
-                chain_ids=[x + 1 for x in range(chains)],
-                method_args=sampler_args,
-            )
-            runset = RunSet(args=cmdstan_args, chains=chains)
-            runset._csv_files = csvfiles
-            for i in range(len(runset._retcodes)):
-                runset._set_retcode(i, 0)
-            fit = CmdStanMCMC(runset)
-            fit.draws()
-            return fit
-        elif config_dict['method'] == 'optimize':
-            if 'algorithm' not in config_dict:
-                raise ValueError(
-                    "Cannot find optimization algorithm"
-                    " in file {}.".format(csvfiles[0])
-                )
-            optimize_args = OptimizeArgs(
-                algorithm=config_dict['algorithm'],
-                save_iterations=config_dict['save_iterations'],
-            )
-            cmdstan_args = CmdStanArgs(
-                model_name=config_dict['model'],
-                model_exe=config_dict['model'],
-                chain_ids=None,
-                method_args=optimize_args,
-            )
-            runset = RunSet(args=cmdstan_args)
-            runset._csv_files = csvfiles
-            for i in range(len(runset._retcodes)):
-                runset._set_retcode(i, 0)
-            return CmdStanMLE(runset)
-        elif config_dict['method'] == 'variational':
-            if 'algorithm' not in config_dict:
-                raise ValueError(
-                    "Cannot find variational algorithm"
-                    " in file {}.".format(csvfiles[0])
-                )
-            variational_args = VariationalArgs(
-                algorithm=config_dict['algorithm'],
-                iter=config_dict['iter'],
-                grad_samples=config_dict['grad_samples'],
-                elbo_samples=config_dict['elbo_samples'],
-                eta=config_dict['eta'],
-                tol_rel_obj=config_dict['tol_rel_obj'],
-                eval_elbo=config_dict['eval_elbo'],
-                output_samples=config_dict['output_samples'],
-            )
-            cmdstan_args = CmdStanArgs(
-                model_name=config_dict['model'],
-                model_exe=config_dict['model'],
-                chain_ids=None,
-                method_args=variational_args,
-            )
-            runset = RunSet(args=cmdstan_args)
-            runset._csv_files = csvfiles
-            for i in range(len(runset._retcodes)):
-                runset._set_retcode(i, 0)
-            return CmdStanVB(runset)
-        else:
-            get_logger().info(
-                'Unable to process CSV output files from method %s.',
-                (config_dict['method']),
-            )
-            return None
-    except (IOError, OSError, PermissionError) as e:
-        raise ValueError(
-            'An error occured processing the CSV files:\n\t{}'.format(str(e))
-        ) from e
 
 
 def build_xarray_data(
